reorganises github examples, documents github checks example
